refactor : all functions, add optional

+++ app/services/user_service.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timezone
+from typing import Optional
 
 from app.models.user import User
 from app.schemas.user import UserCreate
@@ -7,31 +8,38 @@ from pydantic import EmailStr
 from sqlmodel import Session, select
 
 
-def create_user(db: Session, user_data: UserCreate) -> User:
+def create_user(db: Session, user_data: UserCreate) -> Optional[User]:
     # if para asegurar que el cliente no exista en la base de datos
     if obtener_usuario_por_email(db, user_data.email):
         raise HTTPException(status_code=400, detail="EL usuario ya existe")
     nuevo_user = User(**user_data.model_dump())
     db.add(nuevo_user)
-    db.commit()
+    try:
+        db.commit()
+    except Exception:
+        db.rollback()
+        raise HTTPException(
+            status_code=400, detail="Error de integridad, intente de nuevo"
+        )
+
     db.refresh(nuevo_user)
 
     return nuevo_user
 
 
-def obtener_usuario_por_id(db: Session, user_id: int):
+def obtener_usuario_por_id(db: Session, user_id: int) -> Optional[User]:
     return db.get(User, user_id)
 
 
-def obtener_usuario_por_email(db: Session, user_email: EmailStr):
+def obtener_usuario_por_email(db: Session, user_email: EmailStr) -> Optional[User]:
     sentencia = select(User).where(User.email == user_email)
-    resultados = db.exec(sentencia).firts()
+    resultados = db.exec(sentencia).first()
     if not resultados:
         raise HTTPException(status_code=404, detail="EL usuario no existe.")
     return resultados
 
 
-def modificar_usuario(db: Session, user_data: User):
+def modificar_usuario(db: Session, user_data: User) -> User:
     user_db = db.get(User, User.id)
     if not user_db:
         raise HTTPException(status_code=400, detail="Usuario no encontrado")
@@ -46,8 +54,10 @@ def modificar_usuario(db: Session, user_data: User):
     return user_db
 
 
-def eliminar_usuario(db: Session, user_id: int):
+def eliminar_usuario(db: Session, user_id: int) -> dict:
     user_db = obtener_usuario_por_id(db, user_id)
+    if not user_db:
+        raise HTTPException(status_code=404, detail="Usuario no encontrado")
     db.delete(user_db)
     db.commit()
     return {"ok": True, "message": "Usuario eliminado"}
